refactor ASM extensions into multiple files
