+++ agent/src/main/kotlin/net/onedaybeard/graftt/agent/GraftAgent.kt
@@ -2,6 +2,8 @@ package net.onedaybeard.graftt.agent
 
 import com.github.michaelbull.result.*
 import net.onedaybeard.graftt.*
+import net.onedaybeard.graftt.asm.classNode
+import net.onedaybeard.graftt.asm.toBytes
 import net.onedaybeard.graftt.graft.isTransplant
 import net.onedaybeard.graftt.graft.transplant
 import net.onedaybeard.graftt.graft.readRecipientType
