+++ core/src/test/kotlin/net/onedaybeard/graftt/AsmTestUtilKt.kt
@@ -1,6 +1,7 @@
 package net.onedaybeard.graftt
 
 import com.github.michaelbull.result.getError
+import net.onedaybeard.graftt.asm.classNode
 import org.objectweb.asm.tree.ClassNode
 import kotlin.reflect.KClass
 import kotlin.test.assertEquals
